refactor addmembermodal tests
